+++ docs_src/connections/asgi.py
@@ -14,4 +14,4 @@ app = models.asgi(
 # load settings
 monkay.evaluate_settings(ignore_import_errors=False)
 # monkey-patch app so you can use edgy shell
-monkay.set_instance(Instance(registry=registry, app=app))
+monkay.set_instance(Instance(registry=models, app=app))