+++ docs_src/connections/asynccontextmanager.py
@@ -7,6 +7,6 @@ async def main():
     # load settings
     monkay.evaluate_settings(ignore_import_errors=False)
     # monkey-patch so you can use edgy shell
-    monkay.set_instance(Instance(registry=registry))
+    monkay.set_instance(Instance(registry=models))
     async with models:
         ...