+++ docs/release-notes.md
@@ -10,6 +10,8 @@ hide:
 
 ### Added
 
+- Allow multiple admins.
+- Add asgi multiplexing middleware (`edgy.contrib.lilya.middleware.EdgyMiddleware`).
 - Support for `isnull` operator for querysets on a field.
 - Support `NumberedPaginator` as alias for `Paginator` (n fact `Paginator` is an alias of `NumberedPaginator`).